refactor: 컨트롤러 관련 설정 ui.config로 이동
