+++ src/main/java/shopping/ui/config/AuthenticationPrincipalArgumentResolver.java
@@ -1,4 +1,4 @@
-package shopping.ui;
+package shopping.ui.config;
 
 import org.springframework.core.MethodParameter;
 import org.springframework.stereotype.Component;
