+++ src/main/java/shopping/ui/config/WebMvcConfiguration.java
@@ -1,12 +1,10 @@
-package shopping.config;
+package shopping.ui.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.ViewControllerRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import shopping.config.interceptor.AuthInterceptor;
-import shopping.ui.AuthenticationPrincipalArgumentResolver;
 
 import java.util.List;
